major compromise for construction
Code generation is a second-best solution in the first place. I'd rather not be overriding constructors at all. With that in mind, this compromise adds a Mapper::getTableClass() static method to report the table class to use when constructing the Mapper. This is certainly not orthodox and definitely not a good practice in general, but it does help avoid boilerplate code generation. The only reason for having it at all is that a Mapper may depend on any Table, not just the one in its own namespace (think single-table inheritance, etc). Not especially proud of this but it's bearable.